Change length of heart rate and SpO2 to 16-bits

+++ wristband_app/Application/max32664_task.c
@@ -72,8 +72,8 @@ typedef enum {
 
 // Heart Rate Data
 typedef struct {
-    uint8_t heartRate;
-    uint8_t spO2;
+    uint16_t heartRate;
+    uint16_t spO2;
     uint8_t confidence;
     uint8_t status;
 } heartrate_data_t;
